+++ packages/framework/next-ui/src/index.ts
@@ -81,5 +81,3 @@ export { IGRPNestedProviders, type IGRPNestedProvidersArgs } from './components/
 export { IGRPRootProviders } from './components/providers/root';
 
 export { IGRPSessionProvider } from './components/providers/session';
-
-